+++ pdfio-content.c
@@ -4513,7 +4513,7 @@ write_string(pdfio_stream_t *st,	// I - Stream
     if (unicode)
     {
       // Write UTF-16 in hex...
-      if (ch < 0x100000)
+      if (ch < 0x10000)
       {
         // Two-byte UTF-16
 	if (!pdfioStreamPrintf(st, "%04X", ch))
@@ -4521,7 +4521,8 @@ write_string(pdfio_stream_t *st,	// I - Stream
       }
       else
       {
-        // Four-byte UTF-16
+        // Four-byte UTF-16 (surrogate pair)
+	ch -= 0x10000;
 	if (!pdfioStreamPrintf(st, "%04X%04X", 0xd800 | ((ch >> 10) & 0x03ff), 0xdc00 | (ch & 0x03ff)))
 	  return (false);
       }