+++ pdfio-value.c
@@ -356,6 +356,8 @@ _pdfioValueRead(pdfio_file_t   *pdf,	// I - PDF file
 
   PDFIO_DEBUG("_pdfioValueRead(pdf=%p, obj=%p, v=%p)\n", (void *)pdf, (void *)obj, (void *)v);
 
+  memset(v, 0, sizeof(_pdfio_value_t));
+
   if (!token)
     goto done;
 
@@ -371,9 +373,10 @@ _pdfioValueRead(pdfio_file_t   *pdf,	// I - PDF file
       goto done;
     }
 
-    v->type = PDFIO_VALTYPE_ARRAY;
     if ((v->value.array = _pdfioArrayRead(pdf, obj, tb, depth + 1)) == NULL)
       goto done;
+
+    v->type = PDFIO_VALTYPE_ARRAY;
 
     ret = v;
   }
@@ -386,9 +389,10 @@ _pdfioValueRead(pdfio_file_t   *pdf,	// I - PDF file
       goto done;
     }
 
-    v->type = PDFIO_VALTYPE_DICT;
     if ((v->value.dict = _pdfioDictRead(pdf, obj, tb, depth + 1)) == NULL)
       goto done;
+
+    v->type = PDFIO_VALTYPE_DICT;
 
     ret = v;
   }
@@ -406,7 +410,7 @@ _pdfioValueRead(pdfio_file_t   *pdf,	// I - PDF file
       // String
       v->type         = PDFIO_VALTYPE_STRING;
       v->value.string = pdfioStringCreate(pdf, token + 1);
-      ret           = v;
+      ret             = v;
     }
   }
   else if (token[0] == '/')
@@ -422,13 +426,14 @@ _pdfioValueRead(pdfio_file_t   *pdf,	// I - PDF file
     const char		*tokptr;	// Pointer into token
     unsigned char	*dataptr;	// Pointer into data
 
-    v->type                 = PDFIO_VALTYPE_BINARY;
     v->value.binary.datalen = strlen(token) / 2;
     if ((v->value.binary.data = (unsigned char *)malloc(v->value.binary.datalen)) == NULL)
     {
       _pdfioFileError(pdf, "Out of memory for hex string.");
       goto done;
     }
+
+    v->type = PDFIO_VALTYPE_BINARY;
 
     // Convert hex to binary...
     tokptr  = token + 1;