+++ CHANGES.md
@@ -9,6 +9,7 @@ v1.6.5 - YYYY-MM-DD
 - Fixed `pdfioContentTextXxx` output with Unicode characters past U+FFFF
   (Issue #171)
 - Fixed `pdfioPageCopy` with a damaged input PDF file (Issue #174)
+- Fixed error recovery when reading values (Issue #178)
 - Fixed `pdfioContentTextXxx` output with invalid UTF-8.
 - Fixed import of 16-bit PNG files with alpha.
 - Fixed `pdfioDictSetStringf` which wasn't making a copy of the formatted